move typings to index.d.t
