+++ app/src/Types/index.d.ts
@@ -1,0 +1,25 @@
+type PlayerName = string
+
+interface Rank {
+  rank: number,
+  playerName: PlayerName
+}
+
+type Ranks = Rank[]
+
+interface LeaderboardEntry {
+  id: number,
+  ranks: Ranks
+}
+
+type LeaderboardData = LeaderboardEntry[]
+
+type RankChange = number | 'new'
+
+interface PlayerData extends Rank {
+  rankChange: RankChange
+}
+
+interface PlayerHistory {
+  [id: string]: number[]
+}
